Upgrade Code Coverage Workflow to improve testing (#45)
## Description

<!--- 
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
-->

This PR enhances the code coverage collection process. Previously, the
code was only compiled for macOS, but the reality is that the packages
are available for multiple platforms. Therefore, the workflows now align
with this and test the conformance of all platforms.

## Type of change

<!--- Please delete options that are not relevant. -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

## Checklist

<!--- Please delete options that are not relevant. -->

- [x] My code follows the code style of this project.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [ ] I have added/updated necessary comments/documentation.

---------

Co-authored-by: brennobemoura <37243584+brennobemoura@users.noreply.github.com>